Move generateISSN to separate hepler file
